Renamed components to maintain consistency with figma designs
