+++ src/components/navbar/Navbar.js
@@ -1,52 +1,22 @@
-import TravelerInfoPreview from "./TravelerInfoPreview";
-
-import { useState } from 'react';
-
 import {
-    Collection,
-    ScrollView
+    Card,
+    Button,
+    Image
   } from "@aws-amplify/ui-react";
 
-function Navbar({ items, setCurrentTraveler }) {
-    const [selectedItemIndex, setSelectedItemIndex] = useState(null)
-
-    const handleItemClick = (item, index) => {
-      if (selectedItemIndex === index ) {
-        setCurrentTraveler(null)
-        setSelectedItemIndex(null);
-      }
-      else {
-        setCurrentTraveler(item)
-        setSelectedItemIndex(index)  
-      }
-    }
-
+function Navbar({ signOut }) {
     return (
-      <ScrollView 
-        width="100%" 
-        height="100%" 
-        maxWidth="580px" 
-        columnStart="1"
-        columnEnd="2"
-        rowStart="2"
-        rowEnd="-1"
-      >
-        <Collection
-          items={items}
-          gap="0px"
+        <Card className="Header"
+          rowStart="1"
+          rowEnd="2"
+          columnStart="1"
+          columnEnd="-1"
         >
-          {/* For each item passed to Navbar, create a card object that contains a
-          TravelerInfo component */}
-          {(item, index) => (
-            <TravelerInfoPreview 
-              item={item} 
-              key={index} 
-              onClick={() => handleItemClick(item, index)}
-              isSelected={index === selectedItemIndex}
-            />
-          )}
-        </Collection>
-      </ScrollView>
+          Header
+          <Image alt="Odyssey Logo" src="src/components/logo.jpg"/>
+          <Button onClick={signOut}>Sign Out</Button>
+
+        </Card>
     )
 }
